requirements.txt
fastapi
uvicorn
groq
pillow
numpy
opencv-python-headless
tensorflow-cpu
ultralytics
langchain
langchain-community
faiss-cpu
pandas
h5py
keras
imageio
pydantic
python-multipart
sentence-transformers